LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-datetime-with-step.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<input type=datetime id=datetime step="302400"/>

<script>
description('Test if step limits available dates in calendar picker on datetime input.');

openPicker(document.getElementById('datetime'), test1);

function test1() {
    debug("Step without min or max.");
    shouldBeEqualToString('availableDatesInCurrentMonth().join(",")', '2012-11-01,2012-11-04,2012-11-08,2012-11-11,2012-11-15,2012-11-18,2012-11-22,2012-11-25,2012-11-29');

    // Close popup
    eventSender.keyDown('\x1B');
    waitUntilClosing(test1AfterClose);
}

function test1AfterClose() {
    document.getElementById('datetime').min = "2012-11-16T12:00Z";
    
    // Reopen popup
    openPicker(document.getElementById('datetime'), test2);
}

function test2() {
    debug("Step when min is set.");
    shouldBeEqualToString('availableDatesInCurrentMonth().join(",")', '2012-11-16,2012-11-20,2012-11-23,2012-11-27,2012-11-30');

    // Close popup
    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClose);
}

function test2AfterClose() {
    document.getElementById('datetime').max = "2012-11-20T00:00Z";
    
    // Reopen popup
    openPicker(document.getElementById('datetime'), test3);
}

function test3() {
    debug("Step when min and max are set.");
    shouldBeEqualToString('availableDatesInCurrentMonth().join(",")', '2012-11-16,2012-11-20');
    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
