LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-datetime-with-step-expected.txt

Test if step limits available dates in calendar picker on datetime input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Step without min or max.
PASS availableDatesInCurrentMonth().join(",") is "2012-11-01,2012-11-04,2012-11-08,2012-11-11,2012-11-15,2012-11-18,2012-11-22,2012-11-25,2012-11-29"
Step when min is set.
PASS availableDatesInCurrentMonth().join(",") is "2012-11-16,2012-11-20,2012-11-23,2012-11-27,2012-11-30"
Step when min and max are set.
PASS availableDatesInCurrentMonth().join(",") is "2012-11-16,2012-11-20"
PASS successfullyParsed is true

TEST COMPLETE
 
